+++ src/wallet/FunSmartWallet.ts
@@ -1,13 +1,14 @@
 /* eslint-disable */
-import type { Address } from "viem";
+import type { Address, Hex } from "viem";
 import { UserOperationResponse, UserOperationReceipt } from "@alchemy/aa-core";
 import { SmartWallet } from "./SmartWallet";
+import type { ISmartWallet } from "./ISmartWallet";
 
-export class FunSmartWallet extends SmartWallet {
+export class FunSmartWallet extends SmartWallet implements ISmartWallet {
   async sendUserOperation(
-    _targetAddress: Address,
-    _data: Address,
-    _sponsorGas: boolean
+    target: Address,
+    data: Hex,
+    value?: bigint,
   ): Promise<{ hash: string }> {
     throw new Error("Method not implemented.");
   }
@@ -17,19 +18,19 @@ export class FunSmartWallet extends SmartWallet {
   }
 
   async waitForUserOperationTransaction(
-    _hash: `0x${string}`
+    _hash: `0x${string}`,
   ): Promise<`0x${string}`> {
     throw new Error("Method not implemented.");
   }
 
   async getUserOperationByHash(
-    _hash: `0x${string}`
+    _hash: `0x${string}`,
   ): Promise<UserOperationResponse> {
     throw new Error("Method not implemented.");
   }
 
   async getUserOperationReceipt(
-    _hash: `0x${string}`
+    _hash: `0x${string}`,
   ): Promise<UserOperationReceipt> {
     throw new Error("Method not implemented.");
   }
